requirements.txt
dji_thermal_sdk
matplotlib
numpy
pillow 
torch 
torchvision 
fastapi 
requests 
opencv-python 
ultralytics 
uvicorn 
urllib3==1.26